+++ src/constants/textResponses.ts
@@ -96,6 +96,7 @@ export const phrases: ResponseRepository = [
 	// I18N
 	`Can _you_ speak in ${locales.length} different languages? I didn't think so :p`,
 	`Can you speak more than ${locales.length} different languages? Pretty great if you can, I was just wondering ^^`,
+	[`Can _you_ speak ${locales.length} languages too?`, "Just curious ^^"],
 	() =>
 		`I think I speak "${
 			metadataForLocale(randomSupportedLocale()).nickname
@@ -369,7 +370,7 @@ export const phrases: ResponseRepository = [
 
 	...philosophy,
 	...copypasta
-]; // 228 of these
+]; // 229 of these
 logger.silly(`I have ${phrases.length} random things to say ^^`);
 
 /**