+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- We now verify on startup that all commands were deployed.
+
 ## [2.0.4] - 2023-02-08
 ### Changed
 - Clarified the header to the results from `/test`.
@@ -396,6 +400,7 @@ After updating, be sure to run `npm ci && npm run build:clean && npm run migrate
 ### Added
 - Initial commit
 
+[Unreleased]: https://github.com/AverageHelper/Gamgee/compare/v2.0.4...HEAD
 [2.0.4]: https://github.com/AverageHelper/Gamgee/compare/v2.0.3...v2.0.4
 [2.0.3]: https://github.com/AverageHelper/Gamgee/compare/v2.0.2...v2.0.3
 [2.0.2]: https://github.com/AverageHelper/Gamgee/compare/v2.0.1...v2.0.2